+++ src/largest_cube/mod.rs
@@ -3,6 +3,6 @@ pub mod mapping;
 
 #[derive(PartialEq, Debug, Clone)]
 pub struct LargestCube {
-    pub side_length: usize,
+    pub side_length: u16,
     pub indexes: (usize, usize, usize),
 }
